rafactor(user): Change the login API configuration location
- login query -> command로 변경
